// cms-backup/projects/DigitPop-Player/src/app/models/campaign.ts
export class Campaign {
  name: string;
  budgetType: string;
  budgetAmount: string;
  audienceId: string;
  verificationQuestion: string;
  verificationAnswer: string;
  verificationWrongAnswer1: string;
  verificationWrongAnswer2: string;
  verificationWrongAnswer3: string;
  verificationWrongAnswer4: string;
}